core functions moved to `build_database.py`
